Webservice
These are the PHP scripts that run on the server. They receive the Basic
Name Value Pair parameters as $_POST parameters. The data that is sent
back is in JSON format. Please note that if some servers that you test
these scripts on, they can enable script bots that will add HTML markup
to the data being sent if they do not allow JSON support.